requirements.txt
beautifulsoup4==4.12.3
black==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configparser==7.1.0
idna==3.10
logging==0.4.9.6
mypy-extensions==1.0.0
packaging==24.2
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.3.0
